+++ internal/security_utils/config.go
@@ -4,8 +4,8 @@
 package security_utils
 
 const (
-	CollectorVersion = "1.4.0"
+	CollectorVersion = "1.5.0"
 	JsonVersion      = "1.2.9"
 	CollectorType    = "GOLANG"
-	BuildNumber      = "160"
+	BuildNumber      = "161"
 )