+++ security_handlers/fuzz_request_handler.go
@@ -19,7 +19,10 @@ const (
 	maxPoolSize = 3
 )
 
-var FuzzHandler = RestRequestThreadPool{fuzzedApi: sync.Map{}}
+var (
+	FuzzHandler    = RestRequestThreadPool{fuzzedApi: sync.Map{}}
+	fuzzcontroller = false
+)
 
 type FuzzTask struct {
 	fuzzRequrestHandler *FuzzRequrestHandler
@@ -27,10 +30,17 @@ type FuzzTask struct {
 	requestID           string
 }
 
+var firstIastReplay = true
+
 func (fTask *FuzzTask) Run() {
 
 	if !secConfig.SecureWS.GetStatus() {
 		logger.Infoln("WS not connected drop FuzzTask ")
+	}
+
+	if firstIastReplay {
+		secConfig.GlobalInfo.ApplicationInfo.SetScanStartTime(time.Now())
+		firstIastReplay = false
 	}
 
 	if fTask.fuzzRequrestHandler.IsGRPC {
@@ -92,15 +102,26 @@ func initRestRequestThreadPool() {
 	FuzzHandler.threadPool = threadpool.NewThreadPool(queueSize, maxPoolSize, logger, "RestRequestThreadPool")
 }
 
-func InitFuzzScheduler() {
-	if !secConfig.GlobalInfo.IsIASTEnable() {
+func CloseFuzzScheduler() {
+	fuzzcontroller = true
+}
+
+func (r *RestRequestThreadPool) initFuzzScheduler() {
+	if !secConfig.GlobalInfo.IsIastMode() {
+		r.isFuzzSchedulerInitialized = false
 		return
 	}
 	if FuzzHandler.threadPool == nil {
 		initRestRequestThreadPool()
 	}
 	eventGeneration.SendLogMessage("initializing fuzz request scheduler", "InitFuzzScheduler", "INFO")
+	fuzzcontroller = false
 	for {
+		if fuzzcontroller {
+			r.isFuzzSchedulerInitialized = false
+			logger.Info("WebSocket Fuzz scheduler thread stopped")
+			return
+		}
 		iastProbingInterval := secConfig.GlobalInfo.IastProbingInterval()
 		logger.Debugln("iastProbingInterval SleepTime", iastProbingInterval)
 		time.Sleep(time.Duration(iastProbingInterval) * time.Second)
@@ -121,16 +142,21 @@ func InitFuzzScheduler() {
 			//logger.Debugln("LastFuzzRequestTime", FuzzHandler.LastFuzzRequestTime(), currentTime)
 			continue
 		}
+		currentFetchThreshold := secConfig.GlobalInfo.ScanControllersIastLoadInterval() / 12
 
-		currentFetchThreshold := 300 //SECURITY_POLICY_VULNERABILITY_SCAN_IAST_SCAN_PROBING_THRESHOLD
+		if currentFetchThreshold <= 0 {
+			return
+		}
+
+		fetchRatio := 300 / currentFetchThreshold
 		remainingRecordCapacity := FuzzHandler.threadPool.RemainingCapacity()
 		currentRecordBacklog := FuzzHandler.threadPool.PendingTask()
 		batchSize := currentFetchThreshold - currentRecordBacklog
 		logger.Debugln("InitFuzzScheduler test ", batchSize, remainingRecordCapacity, currentRecordBacklog, currentFetchThreshold)
 
-		if batchSize > 100 && remainingRecordCapacity > batchSize {
-			logger.Debugln("InitFuzzScheduler", batchSize*2)
-			eventGeneration.IASTDataRequest(batchSize*2, FuzzHandler.CompletedRequestIds(), FuzzHandler.PendingRequestIds())
+		if batchSize > 100/fetchRatio && remainingRecordCapacity > batchSize {
+			logger.Debugln("InitFuzzScheduler", batchSize)
+			eventGeneration.IASTDataRequest(batchSize, FuzzHandler.CompletedRequestIds(), FuzzHandler.PendingRequestIds())
 		}
 	}
 }