+++ security_intercept/intercept.go
@@ -86,6 +86,9 @@ func HookWrapRawNamed(from string, to, toc interface{}) (string, error) {
  */
 
 func TraceFileOperation(fname string, flag int, isFileOpen bool) *secUtils.EventTracker {
+	if secConfig.GlobalInfo.IsInvalidFileAccessDisabled() {
+		return nil
+	}
 	securityHomePath := secConfig.GlobalInfo.SecurityHomePath()
 	if securityHomePath != "" && secUtils.CaseInsensitiveContains(fname, filepath.Join(securityHomePath, "nr-security-home", "logs")) {
 		// here dont put logger, will cause issue with logrus and hook
@@ -118,7 +121,7 @@ func TraceFileOperation(fname string, flag int, isFileOpen bool) *secUtils.Event
  */
 
 func TraceSystemCommand(command string) *secUtils.EventTracker {
-	if !isAgentInitialized() || command == "" {
+	if !isAgentInitialized() || command == "" || secConfig.GlobalInfo.IsCommandInjectionDisabled() {
 		return nil
 	}
 	var arg []string
@@ -138,7 +141,7 @@ func TraceMongoHooks(e error) {
 }
 
 func TraceMongoOperation(arguments []byte, queryType string) *secUtils.EventTracker {
-	if !isAgentInitialized() {
+	if !isAgentInitialized() || secConfig.GlobalInfo.IsNosqlInjectionDisabled() {
 		return nil
 	}
 	if len(arguments) == 0 {
@@ -174,7 +177,7 @@ func TraceSqlHooks(e error) {
 }
 
 func TraceSqlOperation(query string, args ...interface{}) *secUtils.EventTracker {
-	if !isAgentInitialized() {
+	if !isAgentInitialized() || secConfig.GlobalInfo.IsSQLInjectionDisabled() {
 		return nil
 	}
 	if query == "" {
@@ -196,7 +199,7 @@ func TraceSqlOperation(query string, args ...interface{}) *secUtils.EventTracker
 }
 
 func TracePrepareStatement(q, p string) {
-	if !isAgentInitialized() {
+	if !isAgentInitialized() || secConfig.GlobalInfo.IsSQLInjectionDisabled() {
 		return
 	}
 	if q == "" {
@@ -206,7 +209,7 @@ func TracePrepareStatement(q, p string) {
 }
 
 func TraceExecPrepareStatement(q_address string, args ...interface{}) *secUtils.EventTracker {
-	if !isAgentInitialized() {
+	if !isAgentInitialized() || secConfig.GlobalInfo.IsSQLInjectionDisabled() {
 		return nil
 	}
 	parameters := map[int]interface{}{}
@@ -223,7 +226,7 @@ func TraceExecPrepareStatement(q_address string, args ...interface{}) *secUtils.
  */
 
 func TraceXpathOperation(a string) *secUtils.EventTracker {
-	if !isAgentInitialized() || a == "" {
+	if !isAgentInitialized() || a == "" || secConfig.GlobalInfo.IsXpathInjectionDisabled() {
 		return nil
 	}
 	var arg []string
@@ -236,7 +239,7 @@ func TraceXpathOperation(a string) *secUtils.EventTracker {
  */
 
 func TraceLdapOperation(a map[string]string) *secUtils.EventTracker {
-	if !isAgentInitialized() {
+	if !isAgentInitialized() || secConfig.GlobalInfo.IsLdapInjectionDisabled() {
 		return nil
 	}
 	var arg []interface{}
@@ -249,7 +252,7 @@ func TraceLdapOperation(a map[string]string) *secUtils.EventTracker {
  */
 
 func TraceJsOperation(a string) *secUtils.EventTracker {
-	if !isAgentInitialized() || a == "" {
+	if !isAgentInitialized() || a == "" || secConfig.GlobalInfo.IsJavascriptInjectionDisabled() {
 		return nil
 	}
 	var arg []string
@@ -290,13 +293,24 @@ func AssociateApplicationPort(data string) {
 
 // TraceIncommingRequest - interception of incoming request
 
-func TraceIncommingRequest(url, host string, hdrMap map[string][]string, method string, body string, queryparam map[string][]string, protocol, serverName, reqtype string, bodyReader secUtils.SecWriter, csecAttributes map[string]any) {
+func TraceIncommingRequest(url, host string, hdrMap map[string][]string, method string, body string, queryparam map[string][]string, protocol, serverName, reqtype string, bodyReader secUtils.SecWriter, csecAttributes map[string]any, traceId string) {
 	if !isAgentInitialized() {
+		return
+	}
+	route := ""
+	for k, v := range csecAttributes {
+		if secUtils.CaseInsensitiveEquals(k, AttributeCsecRoute) {
+			route, _ = v.(string)
+		}
+	}
+
+	if ok, _ := isSkipIastScanApi(url, route); ok {
 		return
 	}
 
 	infoReq := new(secUtils.Info_req)
 	infoReq.Request.URL = url
+	infoReq.Request.Route = route
 	infoReq.Request.ParameterMap = queryparam
 	infoReq.Request.ClientIP, infoReq.Request.ClientPort = getIpAndPort(host)
 	infoReq.Request.ServerPort = getServerPort()
@@ -311,15 +325,11 @@ func TraceIncommingRequest(url, host string, hdrMap map[string][]string, method 
 	infoReq.Request.ServerName = serverName
 	infoReq.BodyLimit = secConfig.GlobalInfo.BodyLimit()
 	infoReq.ParentID = getHeaderValue(hdrMap, NR_CSEC_PARENT_ID)
+	infoReq.Request.URI = getRequestUri(url)
+	infoReq.TraceId = traceId
 
 	if reqtype == "gRPC" {
 		infoReq.Request.IsGRPC = true
-	}
-
-	for k, v := range csecAttributes {
-		if secUtils.CaseInsensitiveEquals(k, AttributeCsecRoute) {
-			infoReq.Request.Route, _ = v.(string)
-		}
 	}
 
 	secConfig.Secure.AssociateInboundRequest(infoReq)
@@ -448,6 +458,10 @@ func traceResponseOperations() {
 }
 
 func checkSecureCookies(responseHeader http.Header) {
+
+	if secConfig.GlobalInfo.IsInsecureSettingsDisabled() {
+		return
+	}
 	if responseHeader != nil {
 		logger.Debugln("Verifying SecureCookies, response header", responseHeader)
 		tmpRes := http.Response{Header: responseHeader}
@@ -470,7 +484,7 @@ func checkSecureCookies(responseHeader http.Header) {
 }
 
 func xssCheck(r *secUtils.Info_req) {
-	if IsRxssEnabled() && r.ResponseBody != "" {
+	if !IsRxssDisabled() && r.ResponseBody != "" {
 
 		contentType := r.ResponseContentType
 		if !secUtils.IsContentTypeSupported(contentType) {
@@ -565,7 +579,7 @@ func SendExitEvent(exitEvent interface{}, err error) {
 func ProcessInit(server_name string) {
 	secConfig.GlobalInfo.ApplicationInfo.SetServerName(server_name)
 
-	if secConfig.GlobalInfo.MetaData.GetAccountID() != "" {
+	if secConfig.GlobalInfo.MetaData.GetAccountID() != "" && secConfig.SecureWS != nil {
 		eventGeneration.SendApplicationInfo()
 	}
 }
@@ -601,7 +615,7 @@ func UpdateLinkData(linkingMetadata map[string]string) {
 			secConfig.GlobalInfo.MetaData.SetEntityName(entityname)
 		}
 		if !IsDisable() && !IsForceDisable() {
-			go secWs.InitializeWsConnecton()
+			go StartWsConnection()
 		}
 	} else {
 		secConfig.GlobalInfo.MetaData.SetLinkingMetadata(linkingMetadata)
@@ -611,7 +625,7 @@ func UpdateLinkData(linkingMetadata map[string]string) {
 		}
 		if secConfig.SecureWS != nil {
 			SendLogMessage("Reconnect security agent at refresh call", "UpdateLinkData", "INFO")
-			secConfig.SecureWS.ReconnectAtAgentRefresh()
+			go secConfig.SecureWS.ReconnectAtAgentRefresh()
 		}
 	}
 
@@ -686,6 +700,10 @@ func SendEvent(caseType string, data ...interface{}) interface{} {
 func inboundcallHandler(data ...interface{}) {
 
 	csecAttributes := map[string]any{}
+	traceId := ""
+	if len(data) >= 3 {
+		traceId, _ = data[2].(string)
+	}
 	if len(data) >= 2 {
 		csecAttributes, _ = data[1].(map[string]any)
 	}
@@ -696,7 +714,7 @@ func inboundcallHandler(data ...interface{}) {
 
 	r, ok := request.(webRequestv2)
 	if !ok || r == nil {
-		inboundcallHandlerv1(request, csecAttributes)
+		inboundcallHandlerv1(request, csecAttributes, traceId)
 		return
 	}
 	queryparam := map[string][]string{}
@@ -709,11 +727,11 @@ func inboundcallHandler(data ...interface{}) {
 	}
 
 	reqBodyWriter := secUtils.SecWriter{GetBody: r.GetBody, IsDataTruncated: r.IsDataTruncated}
-	TraceIncommingRequest(r.GetURL().String(), clientHost, r.GetHeader(), r.GetMethod(), "", queryparam, r.GetTransport(), r.GetServerName(), r.Type1(), reqBodyWriter, csecAttributes)
+	TraceIncommingRequest(r.GetURL().String(), clientHost, r.GetHeader(), r.GetMethod(), "", queryparam, r.GetTransport(), r.GetServerName(), r.Type1(), reqBodyWriter, csecAttributes, traceId)
 }
 
 // merge inboundcallHandler and inboundcallHandlerv1 in the next major release(v1.0.0)
-func inboundcallHandlerv1(request interface{}, csecAttributes map[string]any) {
+func inboundcallHandlerv1(request interface{}, csecAttributes map[string]any, traceId string) {
 	r, ok := request.(webRequest)
 	if !ok || r == nil {
 		SendLogMessage("ERROR: Request is not a type of webRequest and webRequestv2 ", "security_intercept", "SEVERE")
@@ -730,11 +748,11 @@ func inboundcallHandlerv1(request interface{}, csecAttributes map[string]any) {
 	}
 
 	reqBodyWriter := secUtils.SecWriter{GetBody: r.GetBody, IsDataTruncated: IsDataTruncated}
-	TraceIncommingRequest(r.GetURL().String(), clientHost, r.GetHeader(), r.GetMethod(), "", queryparam, r.GetTransport(), r.GetServerName(), r.Type1(), reqBodyWriter, csecAttributes)
+	TraceIncommingRequest(r.GetURL().String(), clientHost, r.GetHeader(), r.GetMethod(), "", queryparam, r.GetTransport(), r.GetServerName(), r.Type1(), reqBodyWriter, csecAttributes, traceId)
 }
 
 func outboundcallHandler(req interface{}) *secUtils.EventTracker {
-	if req == nil || !isAgentInitialized() {
+	if req == nil || !isAgentInitialized() || secConfig.GlobalInfo.IsSsrfDisabled() {
 		return nil
 	}
 	r, ok := req.(*http.Request)
@@ -885,7 +903,7 @@ func associateApplicationPort(data ...interface{}) {
 }
 
 func mongoHandler(data ...interface{}) *secUtils.EventTracker {
-	if !isAgentInitialized() {
+	if !isAgentInitialized() || secConfig.GlobalInfo.IsNosqlInjectionDisabled() {
 		return nil
 	}
 	if len(data) < 2 {